Issue #1045 - Update projection
- Fix warning because of implicit size_t -> double cast.

+++ libosmscout/include/osmscout/projection/Projection.h
@@ -232,8 +232,8 @@ namespace osmscout {
     {
       return {Vertex2D(0.0,
                        0.0),
-              Vertex2D(GetWidth(),
-                       GetHeight())};
+              Vertex2D(static_cast<double>(GetWidth()),
+                       static_cast<double>(GetHeight()))};
     }
 
     /**
